+++ src/main/java/at/porscheinformatik/seleniumcomponents/EditableSeleniumComponent.java
@@ -10,7 +10,7 @@ import org.openqa.selenium.WebElement;
  *
  * @author ham
  */
-public interface EditableSeleniumComponent extends ClickableSeleniumComponent
+public interface EditableSeleniumComponent extends ClickableSeleniumComponent, DeactivateableSeleniumComponent
 {
 
     /**
@@ -27,28 +27,6 @@ public interface EditableSeleniumComponent extends ClickableSeleniumComponent
                 WebElement element = element();
 
                 return element.isDisplayed() && element.isEnabled();
-            });
-        }
-        catch (NoSuchElementException e)
-        {
-            return false;
-        }
-    }
-
-    /**
-     * Returns true if the component is enabled. Waits {@link SeleniumGlobals#getShortTimeoutInSeconds()} seconds for
-     * the component to become available.
-     *
-     * @return true if enabled
-     */
-    default boolean isEnabled()
-    {
-        try
-        {
-            return SeleniumUtils.retryOnStale(() -> {
-                WebElement element = element();
-
-                return element.isEnabled();
             });
         }
         catch (NoSuchElementException e)