+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/ButtonComponent.java
@@ -1,5 +1,5 @@
 /**
- * 
+ *
  */
 package at.porscheinformatik.seleniumcomponents.component;
 
@@ -7,13 +7,15 @@ import static at.porscheinformatik.seleniumcomponents.WebElementSelector.*;
 
 import at.porscheinformatik.seleniumcomponents.AbstractSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.ClickableSeleniumComponent;
+import at.porscheinformatik.seleniumcomponents.DeactivateableSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 
 /**
  * @author Daniel Furtlehner
  */
-public class ButtonComponent extends AbstractSeleniumComponent implements ClickableSeleniumComponent
+public class ButtonComponent extends AbstractSeleniumComponent
+    implements ClickableSeleniumComponent, DeactivateableSeleniumComponent
 {
     public ButtonComponent(SeleniumComponent parent)
     {