+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/CheckboxComponent.java
@@ -6,6 +6,7 @@ package at.porscheinformatik.seleniumcomponents.component;
 import static at.porscheinformatik.seleniumcomponents.WebElementSelector.*;
 
 import at.porscheinformatik.seleniumcomponents.AbstractSeleniumComponent;
+import at.porscheinformatik.seleniumcomponents.DeactivateableSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SelectableSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
@@ -13,7 +14,8 @@ import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 /**
  * @author Daniel Furtlehner
  */
-public class CheckboxComponent extends AbstractSeleniumComponent implements SelectableSeleniumComponent
+public class CheckboxComponent extends AbstractSeleniumComponent
+    implements SelectableSeleniumComponent, DeactivateableSeleniumComponent
 {
 
     /**
@@ -30,18 +32,6 @@ public class CheckboxComponent extends AbstractSeleniumComponent implements Sele
     public CheckboxComponent(SeleniumComponent parent)
     {
         this(parent, selectByCss("input[type=\"checkbox\"]"));
-    }
-
-    public void setSelected(boolean selected)
-    {
-        if (selected)
-        {
-            select();
-        }
-        else
-        {
-            unselect();
-        }
     }
 
     public String getId()