+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/RadioComponent.java
@@ -6,6 +6,7 @@ package at.porscheinformatik.seleniumcomponents.component;
 import static at.porscheinformatik.seleniumcomponents.WebElementSelector.*;
 
 import at.porscheinformatik.seleniumcomponents.AbstractSeleniumComponent;
+import at.porscheinformatik.seleniumcomponents.DeactivateableSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SelectableSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
@@ -13,7 +14,8 @@ import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 /**
  * @author Daniel Furtlehner
  */
-public class RadioComponent extends AbstractSeleniumComponent implements SelectableSeleniumComponent
+public class RadioComponent extends AbstractSeleniumComponent
+    implements SelectableSeleniumComponent, DeactivateableSeleniumComponent
 {
 
     /**
@@ -31,17 +33,4 @@ public class RadioComponent extends AbstractSeleniumComponent implements Selecta
     {
         this(parent, selectByCss("input[type=\"radio\"]"));
     }
-
-    public void setSelected(boolean selected)
-    {
-        if (selected)
-        {
-            select();
-        }
-        else
-        {
-            unselect();
-        }
-    }
-
 }